pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Django project with REST API"
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.1",
    "django-filter>=24.2",
    "drf-spectacular>=0.27.2",
    "django-extensions>=3.2.3",
    "pygraphviz>=1.12",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"